+++ flit/__init__.py
@@ -7,7 +7,7 @@ import sys
 from . import common
 from .log import enable_colourful_output
 
-__version__ = '0.7.2'
+__version__ = '0.11.4'
 
 log = logging.getLogger(__name__)
 
@@ -18,12 +18,15 @@ def add_shared_install_options(parser):
     parser.add_argument('--env', action='store_false', dest='user',
         help="Install into sys.prefix (default if site.ENABLE_USER_SITE is False, i.e. in virtualenvs)"
     )
+    parser.add_argument('--python', default=sys.executable,
+        help="Target Python executable, if different from the one running flit"
+    )
 
 def main(argv=None):
     ap = argparse.ArgumentParser()
     ap.add_argument('-f', '--ini-file', type=pathlib.Path, default='flit.ini')
     ap.add_argument('--version', action='version', version='Flit '+__version__)
-    ap.add_argument('--repository', default='pypi',
+    ap.add_argument('--repository',
         help="Name of the repository to upload to (must be in ~/.pypirc)"
     )
     ap.add_argument('--debug', action='store_true', help=argparse.SUPPRESS)
@@ -38,6 +41,26 @@ def main(argv=None):
     )
     parser_wheel.add_argument('--verify-metadata', action='store_true',
           help="Verify the package metadata with the PyPI server"
+    )
+
+    parser_sdist = subparsers.add_parser('sdist',
+         help="Build a source distribution (.tar.gz)",
+    )
+
+    parser_build = subparsers.add_parser('build',
+        help="Build wheel and sdist",
+    )
+
+    parser_build.add_argument('--format', action='append',
+        help="Select a format to build. Options: 'wheel', 'sdist'"
+    )
+
+    parser_publish = subparsers.add_parser('publish',
+        help="Upload wheel and sdist",
+    )
+
+    parser_publish.add_argument('--format', action='append',
+        help="Select a format to publish. Options: 'wheel', 'sdist'"
     )
 
     parser_install = subparsers.add_parser('install',
@@ -78,22 +101,37 @@ def main(argv=None):
         sys.exit(0)
 
     if args.subcmd == 'wheel':
-        from .wheel import WheelBuilder
+        from .wheel import wheel_main
         try:
-            WheelBuilder(args.ini_file, upload=args.upload,
+            wheel_main(args.ini_file, upload=args.upload,
                      verify_metadata=args.verify_metadata,
-                     repo=args.repository).build()
+                     repo=args.repository)
         except common.ProblemInModule as e:
             sys.exit(e.args[0])
+    elif args.subcmd == 'sdist':
+        from .sdist import SdistBuilder
+        try:
+            SdistBuilder(args.ini_file).build()
+        except common.VCSError as e:
+            sys.exit(str(e))
+
+    elif args.subcmd == 'build':
+        from .build import main
+        main(args.ini_file, formats=set(args.format or []))
+    elif args.subcmd == 'publish':
+        from .upload import main
+        main(args.ini_file, args.repository, formats=set(args.format or []))
+
     elif args.subcmd == 'install':
         from .install import Installer
         try:
-            Installer(args.ini_file, user=args.user, symlink=args.symlink, deps=args.deps).install()
+            Installer(args.ini_file, user=args.user, python=args.python,
+                      symlink=args.symlink, deps=args.deps).install()
         except (common.NoDocstringError, common.NoVersionError) as e:
             sys.exit(e.args[0])
     elif args.subcmd == 'installfrom':
         from .installfrom import installfrom
-        sys.exit(installfrom(args.location, user=args.user))
+        sys.exit(installfrom(args.location, user=args.user, python=args.python))
     elif args.subcmd == 'register':
         from .upload import register
         meta, mod = common.metadata_and_module_from_ini_path(args.ini_file)
@@ -104,6 +142,3 @@ def main(argv=None):
     else:
         ap.print_help()
         sys.exit(1)
-
-if __name__ == '__main__':
-    main()