+++ flit/installfrom.py
@@ -1,9 +1,7 @@
 import os.path
 import pathlib
 import re
-import shutil
 import site
-import subprocess
 import sysconfig
 import sys
 import tarfile
@@ -26,14 +24,14 @@ class InvalidAddress(BadInput):
     def __init__(self, address):
         self.address = address
 
-    def __str__(self):
+    def __str__(self):  # pragma: no cover
         return "Invalid address: {!r}".format(self.address)
 
 class UnknownAddressType(BadInput):
     def __init__(self, address_type):
         self.address_type = address_type
 
-    def __str__(self):
+    def __str__(self):  # pragma: no cover
         return "Unknown address type: {}".format(self.address_type)
 
 class InvalidAddressLocation(BadInput):
@@ -42,7 +40,7 @@ class InvalidAddressLocation(BadInput):
         self.location = location
         self.expected_pattern = expected_pattern
 
-    def __str__(self):
+    def __str__(self):  # pragma: no cover
         return "Invalid location: {!r}\n{}: addresses should look like {}".format(
             self.location, self.address_type, self.expected_pattern
         )
@@ -110,17 +108,19 @@ def fetch(address_type, location):
         return download_unpack(url)
 
 
-def install_local(path, user=False):
+def install_local(path, user=False, python=sys.executable):
     p = pathlib.Path(path)
-    Installer(p / 'flit.ini', user=user, deps='production').install()
+    Installer(p / 'flit.ini', user=user, python=sys.executable,
+              deps='production').install()
 
 
-def installfrom(address, user=None):
+def installfrom(address, user=None, python=sys.executable):
     if user is None:
-        user = site.ENABLE_USER_SITE and not os.access(sysconfig.get_path('purelib'), os.W_OK)
+        user = site.ENABLE_USER_SITE \
+               and not os.access(sysconfig.get_path('purelib'), os.W_OK)
 
     try:
-        return install_local(fetch(*parse_address(address)), user=user)
+        return install_local(fetch(*parse_address(address)), user=user, python=python)
     except BadInput as e:
         print(e, file=sys.stderr)
         return 2