+++ flit/wheel.py
@@ -1,3 +1,4 @@
+from base64 import urlsafe_b64encode
 import configparser
 import contextlib
 from datetime import datetime
@@ -6,7 +7,14 @@ import io
 import logging
 import os
 import re
-import zipfile
+import sys
+import tempfile
+from types import SimpleNamespace
+
+if sys.version_info >= (3, 6):
+    import zipfile
+else:
+    import zipfile36 as zipfile
 
 from flit import __version__
 from . import common
@@ -26,23 +34,15 @@ class EntryPointsConflict(ValueError):
             'flit.ini, not both.')
 
 class WheelBuilder:
-    def __init__(self, ini_path, upload=False, verify_metadata=False, repo='pypi'):
+    def __init__(self, ini_path, target_fp):
         """Build a wheel from a module/package
         """
         self.ini_path = ini_path
         self.directory = ini_path.parent
-        self.dist_dir = self.directory / 'dist'
-        try:
-            self.dist_dir.mkdir()
-        except FileExistsError:
-            pass
 
         self.ini_info = inifile.read_pkg_ini(ini_path)
         self.module = common.Module(self.ini_info['module'], ini_path.parent)
         self.metadata = common.make_metadata(self.module, self.ini_info)
-        self.upload=upload
-        self.verify_metadata=verify_metadata
-        self.repo = repo
 
         self.dist_version = self.metadata.name + '-' + self.metadata.version
         self.records = []
@@ -52,22 +52,22 @@ class WheelBuilder:
             d = datetime.utcfromtimestamp(int(os.environ['SOURCE_DATE_EPOCH']))
             log.info("Zip timestamps will be from SOURCE_DATE_EPOCH: %s", d)
             # zipfile expects a 6-tuple, not a datetime object
-            self.source_time_stamp = (d.year, d.minute, d.day, d.hour, d.minute, d.second)
+            self.source_time_stamp = (d.year, d.month, d.day, d.hour, d.minute, d.second)
         except (KeyError, ValueError):
             # Otherwise, we'll use the mtime of files, and generated files will
             # default to 2016-1-1 00:00:00
             self.source_time_stamp = None
 
         # Open the zip file ready to write
-        self.wheel_zip = zipfile.ZipFile(str(self.wheel_path), 'w',
+        self.wheel_zip = zipfile.ZipFile(target_fp, 'w',
                              compression=zipfile.ZIP_DEFLATED)
 
     @property
-    def wheel_path(self):
+    def wheel_filename(self):
         tag = ('py2.' if self.supports_py2 else '') + 'py3-none-any'
-        return self.dist_dir / '{}-{}-{}.whl'.format(
-                re.sub("[^\w\d.]+", "_", self.metadata.name, re.UNICODE),
-                re.sub("[^\w\d.]+", "_", self.metadata.version, re.UNICODE),
+        return '{}-{}-{}.whl'.format(
+                re.sub("[^\w\d.]+", "_", self.metadata.name, flags=re.UNICODE),
+                re.sub("[^\w\d.]+", "_", self.metadata.version, flags=re.UNICODE),
                 tag)
 
     def _include(self, path):
@@ -79,10 +79,37 @@ class WheelBuilder:
     def _add_file(self, full_path, rel_path):
         log.debug("Adding %s to zip file", full_path)
         full_path, rel_path = str(full_path), str(rel_path)
-        hashsum = my_zip_write(self.wheel_zip, full_path, rel_path,
-                               date_time=self.source_time_stamp)
+        if os.sep != '/':
+            # We always want to have /-separated paths in the zip file and in
+            # RECORD
+            rel_path = rel_path.replace(os.sep, '/')
+
+        if self.source_time_stamp is None:
+            zinfo = zipfile.ZipInfo.from_file(full_path, rel_path)
+        else:
+            # Set timestamps in zipfile for reproducible build
+            zinfo = zipfile.ZipInfo(rel_path, self.source_time_stamp)
+        
+        # Normalize permission bits to either 755 (executable) or 644 (not executable)
+        st_mode = os.stat(full_path).st_mode
+        if st_mode & 0o100:
+            st_mode = (st_mode | 0o755) & ~0o22
+        else:
+            st_mode = (st_mode | 0o644) & ~0o133
+        zinfo.external_attr = (st_mode & 0xFFFF) << 16      # Unix attributes
+
+        hashsum = hashlib.sha256()
+        with open(full_path, 'rb') as src, self.wheel_zip.open(zinfo, 'w') as dst:
+            while True:
+                buf = src.read(1024 * 8)
+                if not buf:
+                    break
+                hashsum.update(buf)
+                dst.write(buf)
+
         size = os.stat(full_path).st_size
-        self.records.append((rel_path, hashsum.hexdigest(), size))
+        hash_digest = urlsafe_b64encode(hashsum.digest()).decode('ascii').rstrip('=')
+        self.records.append((rel_path, hash_digest, size))
 
     @contextlib.contextmanager
     def _write_to_zip(self, rel_path):
@@ -90,14 +117,19 @@ class WheelBuilder:
         yield sio
 
         log.debug("Writing data to %s in zip file", rel_path)
+        # The default is a fixed timestamp rather than the current time, so
+        # that building a wheel twice on the same computer can automatically
+        # give you the exact same result.
         date_time = self.source_time_stamp or (2016, 1, 1, 0, 0, 0)
         zi = zipfile.ZipInfo(rel_path, date_time)
         b = sio.getvalue().encode('utf-8')
         hashsum = hashlib.sha256(b)
+        hash_digest = urlsafe_b64encode(hashsum.digest()).decode('ascii').rstrip('=')
         self.wheel_zip.writestr(zi, b, compress_type=zipfile.ZIP_DEFLATED)
-        self.records.append((rel_path, hashsum.hexdigest(), len(b)))
+        self.records.append((rel_path, hash_digest, len(b)))
 
     def copy_module(self):
+        log.info('Copying package file(s) from %s', self.module.path)
         if self.module.is_package:
             # Walk the tree and compress it, sorting everything so the order
             # is stable.
@@ -115,9 +147,10 @@ class WheelBuilder:
     @property
     def supports_py2(self):
         return not (self.metadata.requires_python or '')\
-                                    .startswith(('3', '>3', '>=3'))
+                                    .startswith(('3', '>3', '>=3', '~=3'))
 
     def write_metadata(self):
+        log.info('Writing metadata files')
         dist_info = self.dist_version + '.dist-info'
 
         # Write entry points
@@ -138,6 +171,9 @@ class WheelBuilder:
         elif self.ini_info['entry_points_file'] is not None:
             self._add_file(self.ini_info['entry_points_file'],
                            dist_info + '/entry_points.txt')
+        for base in ('COPYING', 'LICENSE'):
+            for path in sorted(self.directory.glob(base + '*')):
+                self._add_file(path, '%s/%s' % (dist_info, path.name))
 
         with self._write_to_zip(dist_info + '/WHEEL') as f:
             f.write(wheel_file_template)
@@ -149,6 +185,7 @@ class WheelBuilder:
             self.metadata.write_metadata_file(f)
 
     def write_record(self):
+        log.info('Writing the record of files')
         # Write a record of the files in the wheel
         with self._write_to_zip(self.dist_version + '.dist-info/RECORD') as f:
             for path, hash, size in self.records:
@@ -156,132 +193,40 @@ class WheelBuilder:
             # RECORD itself is recorded with no hash or size
             f.write(self.dist_version + '.dist-info/RECORD,,\n')
 
-    def post_build(self):
-        if self.verify_metadata:
-            from .upload import verify
-            verify(self.metadata, self.repo)
-
-        if self.upload:
-            from .upload import do_upload
-            do_upload(self.wheel_path, self.metadata, self.repo)
-
     def build(self):
-        self.copy_module()
-        self.write_metadata()
-        self.write_record()
+        try:
+            self.copy_module()
+            self.write_metadata()
+            self.write_record()
+        finally:
+            self.wheel_zip.close()
 
-        self.wheel_zip.close()
-
-        self.post_build()
-
-
-import stat, time
-from zipfile import ZipInfo, ZIP_LZMA, _get_compressor, ZIP64_LIMIT, crc32
-
-def my_zip_write(self, filename, arcname=None, compress_type=None,
-                 date_time=None):
-    """Copy of zipfile.ZipFile.write() with some modifications
-
-    - Allow overriding the timestamp for reproducible builds
-    - Calculate a SHA256 hash of the file as we write it and return the hash
-      object.
+def wheel_main(ini_path, upload=False, verify_metadata=False, repo='pypi'):
+    """Build a wheel in the dist/ directory, and optionally upload it.
     """
-    if not self.fp:
-        raise RuntimeError(
-            "Attempt to write to ZIP archive that was already closed")
+    dist_dir = ini_path.parent / 'dist'
+    try:
+        dist_dir.mkdir()
+    except FileExistsError:
+        pass
 
-    st = os.stat(filename)
-    isdir = stat.S_ISDIR(st.st_mode)
-    if date_time is None:
-        mtime = time.localtime(st.st_mtime)
-        date_time = mtime[0:6]
-    # Create ZipInfo instance to store file information
-    if arcname is None:
-        arcname = filename
-    arcname = os.path.normpath(os.path.splitdrive(arcname)[1])
-    while arcname[0] in (os.sep, os.altsep):
-        arcname = arcname[1:]
-    if isdir:
-        arcname += '/'
-    zinfo = ZipInfo(arcname, date_time)
-    st_mode = st.st_mode
-    if st_mode & 0o100:
-        st_mode = (st_mode | 0o755) & ~0o22
-    else:
-        st_mode = (st_mode | 0o644) & ~0o133
-    zinfo.external_attr = (st_mode & 0xFFFF) << 16      # Unix attributes
-    if isdir:
-        zinfo.compress_type = zipfile.ZIP_STORED
-    elif compress_type is None:
-        zinfo.compress_type = self.compression
-    else:
-        zinfo.compress_type = compress_type
+    # We don't know the final filename until metadata is loaded, so write to
+    # a temporary_file, and rename it afterwards.
+    (fd, temp_path) = tempfile.mkstemp(suffix='.whl', dir=str(dist_dir))
+    with open(fd, 'w+b') as fp:
+        wb = WheelBuilder(ini_path, fp)
+        wb.build()
 
-    zinfo.file_size = st.st_size
-    zinfo.flag_bits = 0x00
-    self.fp.seek(getattr(self, 'start_dir', 0))
-    zinfo.header_offset = self.fp.tell()    # Start of header bytes
-    if zinfo.compress_type == ZIP_LZMA:
-        # Compressed data includes an end-of-stream (EOS) marker
-        zinfo.flag_bits |= 0x02
+    wheel_path = dist_dir / wb.wheel_filename
+    os.replace(temp_path, str(wheel_path))
+    log.info("Built wheel: %s", wheel_path)
 
-    self._writecheck(zinfo)
-    self._didModify = True
+    if verify_metadata:
+        from .upload import verify
+        verify(wb.metadata, repo)
 
-    if isdir:
-        zinfo.file_size = 0
-        zinfo.compress_size = 0
-        zinfo.CRC = 0
-        zinfo.external_attr |= 0x10  # MS-DOS directory flag
-        self.filelist.append(zinfo)
-        self.NameToInfo[zinfo.filename] = zinfo
-        self.fp.write(zinfo.FileHeader(False))
-        self.start_dir = self.fp.tell()
-        return
+    if upload:
+        from .upload import do_upload
+        do_upload(wheel_path, wb.metadata, repo)
 
-    hashsum = hashlib.sha256()
-    cmpr = _get_compressor(zinfo.compress_type)
-    with open(filename, "rb") as fp:
-        # Must overwrite CRC and sizes with correct data later
-        zinfo.CRC = CRC = 0
-        zinfo.compress_size = compress_size = 0
-        # Compressed size can be larger than uncompressed size
-        zip64 = self._allowZip64 and \
-            zinfo.file_size * 1.05 > ZIP64_LIMIT
-        self.fp.write(zinfo.FileHeader(zip64))
-        file_size = 0
-        while 1:
-            buf = fp.read(1024 * 8)
-            if not buf:
-                break
-            file_size = file_size + len(buf)
-            CRC = crc32(buf, CRC) & 0xffffffff
-            hashsum.update(buf)
-            if cmpr:
-                buf = cmpr.compress(buf)
-                compress_size = compress_size + len(buf)
-            self.fp.write(buf)
-    if cmpr:
-        buf = cmpr.flush()
-        compress_size = compress_size + len(buf)
-        self.fp.write(buf)
-        zinfo.compress_size = compress_size
-    else:
-        zinfo.compress_size = file_size
-    zinfo.CRC = CRC
-    zinfo.file_size = file_size
-    if not zip64 and self._allowZip64:
-        if file_size > ZIP64_LIMIT:
-            raise RuntimeError('File size has increased during compressing')
-        if compress_size > ZIP64_LIMIT:
-            raise RuntimeError('Compressed size larger than uncompressed size')
-    # Seek backwards and write file header (which will now include
-    # correct CRC and file sizes)
-    self.start_dir = self.fp.tell()       # Preserve current position in file
-    self.fp.seek(zinfo.header_offset, 0)
-    self.fp.write(zinfo.FileHeader(zip64))
-    self.fp.seek(self.start_dir, 0)
-    self.filelist.append(zinfo)
-    self.NameToInfo[zinfo.filename] = zinfo
-
-    return hashsum
+    return SimpleNamespace(builder=wb, file=wheel_path)