+++ tests/test_upload.py
@@ -1,8 +1,10 @@
+from contextlib import contextmanager
+import io
 import pathlib
+import sys
 
-import pytest
-import requests
 import responses
+from testpath import modified_env
 from unittest.mock import patch
 
 from flit import upload, common, wheel
@@ -11,7 +13,8 @@ samples_dir = pathlib.Path(__file__).parent / 'samples'
 
 repo_settings = {'url': upload.PYPI,
                  'username': 'user',
-                 'password': 'pw'
+                 'password': 'pw',
+                 'is_warehouse': True,
                 }
 
 @responses.activate
@@ -38,9 +41,8 @@ def test_verify():
 def test_upload():
     responses.add(responses.POST, upload.PYPI, status=200)
 
-    wb = wheel.WheelBuilder(samples_dir / 'module1-pkg.ini', upload='pypi')
     with patch('flit.upload.get_repository', return_value=repo_settings):
-        wb.build()
+        wheel.wheel_main(samples_dir / 'module1-pkg.ini', upload='pypi')
 
     assert len(responses.calls) == 1
 
@@ -51,12 +53,81 @@ def test_upload_registers():
             status = 200 if register_mock.called else 403
             return (status, {}, '')
 
-        responses.add_callback(responses.POST, upload.PYPI,
+        old_pypi = "https://pypi.python.org/pypi"
+        responses.add_callback(responses.POST, old_pypi,
                                callback=upload_callback)
 
-        wb = wheel.WheelBuilder(samples_dir / 'module1-pkg.ini', upload='pypi')
-        with patch('flit.upload.get_repository', return_value=repo_settings):
-            wb.build()
+        repo = repo_settings.copy()
+        repo['url'] = old_pypi
+        repo['is_warehouse'] = False
+
+        with patch('flit.upload.get_repository', return_value=repo):
+            wheel.wheel_main(samples_dir / 'module1-pkg.ini', upload='pypi')
 
     assert len(responses.calls) == 2
     assert register_mock.call_count == 1
+
+pypirc1 = """
+[distutils]
+index-servers =
+    pypi
+
+[pypi]
+username: fred
+password: s3cret
+"""
+# That's not a real password. Well, hopefully not.
+
+def test_get_repository():
+    repo = upload.get_repository(cfg_file=io.StringIO(pypirc1))
+    assert repo['url'] == upload.PYPI
+    assert repo['username'] == 'fred'
+    assert repo['password'] == 's3cret'
+
+def test_get_repository_env():
+    with modified_env({
+        'FLIT_INDEX_URL': 'https://pypi.example.com',
+        'FLIT_USERNAME': 'alice',
+        'FLIT_PASSWORD': 'p4ssword',  # Also not a real password
+    }):
+        repo = upload.get_repository(cfg_file=io.StringIO(pypirc1))
+        # Because we haven't specified a repo name, environment variables should
+        # have higher priority than the config file.
+        assert repo['url'] == 'https://pypi.example.com'
+        assert repo['username'] == 'alice'
+        assert repo['password'] == 'p4ssword'
+
+@contextmanager
+def _fake_keyring(pw):
+    real_keyring = sys.modules.get('keyring', None)
+    class FakeKeyring:
+        @staticmethod
+        def get_password(service_name, username):
+            return pw
+
+    sys.modules['keyring'] = FakeKeyring()
+
+    try:
+        yield
+    finally:
+        if real_keyring is None:
+            del sys.modules['keyring']
+        else:
+            sys.modules['keyring'] = real_keyring
+
+pypirc2 = """
+[distutils]
+index-servers =
+    pypi
+
+[pypi]
+username: fred
+"""
+
+def test_get_repository_keyring():
+    with modified_env({'FLIT_PASSWORD': None}), \
+            _fake_keyring('tops3cret'):
+        repo = upload.get_repository(cfg_file=io.StringIO(pypirc2))
+
+    assert repo['username'] == 'fred'
+    assert repo['password'] == 'tops3cret'